suppress warnings for Borland 5.82
git-svn-id: http://svn.osdn.net/svnroot/hamigaki@98 c04fc264-1e22-0410-9499-a186082fc69e

+++ hamigaki/trunk/libs/audio/test/aiff_file_test.cpp
@@ -27,7 +27,7 @@ namespace ut = boost::unit_test;
 
 using namespace hamigaki::audio::test;
 
-void aiff_file_test_aux(const audio::pcm_format& fmt, double freq)
+void aiff_file_test_aux(const audio::pcm_format& fmt, float freq)
 {
     std::ostringstream os;
     os
