+++ hamigaki/trunk/libs/audio/test/direct_sound_test.cpp
@@ -25,7 +25,7 @@ namespace ut = boost::unit_test;
 
 using namespace hamigaki::audio::test;
 
-void direct_sound_buffer_test_aux(const audio::pcm_format& fmt, double freq)
+void direct_sound_buffer_test_aux(const audio::pcm_format& fmt, float freq)
 {
     std::ostringstream os;
     os
